requirements.txt
basedpyright==1.31.1
nodejs-wheel-binaries==22.18.0
numpy==2.3.2
pygame-ce==2.5.5
